requirements.txt
opencv_python==4.5.1.48
pandas==1.1.5
tensorflow==2.0.0
Keras==2.3.1
asrtoolkit==0.2.4
matplotlib==3.3.4
streamlit==0.75.0
numpy==1.19.2
Pillow==8.3.1
scikit_learn==0.24.2
